refactor: move login and signup components to their pages
